# ProductionCode/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Elite Construction AI Syndicate API
  description: |
    🏗️ **Elite Construction AI Syndicate - Production API Documentation**
    
    This API provides comprehensive endpoints for:
    - Multi-agent AI orchestration
    - Construction plan analysis with VLM
    - Real-time WebSocket communication
    - LLM chat interfaces
    - Authentication and authorization
    - System monitoring and metrics
    
    **Key Features:**
    - JWT-based authentication
    - WebSocket support for real-time updates
    - VLM (Visual Language Model) plan annotation
    - Multi-target chat routing (agents, LLMs, coordinators)
    - HOAI LP 6/7 compliance checking
    - DIN 276 standard verification
    
  version: 1.0.0
  contact:
    name: Elite Construction AI Syndicate
    email: support@construction-syndicate.ai
  license:
    name: Proprietary
    url: https://construction-syndicate.ai/license
servers:
  - url: https://api.construction-syndicate.ai
    description: Production server
  - url: https://staging.construction-syndicate.ai
    description: Staging server
  - url: http://localhost:3001
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  schemas:
...
    SystemInfo:
      type: object
      properties:
        id:
          type: string
          example: "agent-001"
        name:
          type: string
          example: "Arbitrage Agent Alpha"
        type:
          type: string
          enum: [agent, service, coordinator, llm]
        status:
          type: string
          enum: [active, idle, processing, error, offline]
        metrics:
          type: object
          properties:
            cpu:
              type: number
              format: float
              example: 45.2
            memory:
              type: number
              format: float
              example: 2048.5
            uptime:
              type: integer
              example: 3600000
        capabilities:
          type: array
          items:
            type: string
          example: ["arbitrage", "analysis", "trading"]
...
paths:
...
  /api/systems/{systemId}/status:
    get:
      tags:
        - Systems
      summary: Get system status
      description: Get detailed status of a specific system
      operationId: getSystemStatus
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: systemId
          required: true
          schema:
            type: string
          description: System identifier
      responses:
        '200':
          description: System status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    $ref: '#/components/schemas/SystemInfo'
...